requirements_base.txt
otree-core>=1.1.2
Django==1.8.8 # for heroku, needs to be explicitly in requirements file
git+git://github.com/HealthTechDevelopers/celery@master#celery
